lti-service-python/requirements.txt
# LTI 1.3 Service - Python Implementation
# Requirements file for Python dependencies

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3  # Built-in for development

# JWT and Crypto
pyjwt[crypto]==2.8.0
cryptography==41.0.8
jwcrypto==1.5.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Templates
jinja2==3.1.2

# Environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Date handling
python-dateutil==2.8.2

# CORS
starlette==0.27.0
